+++ News-Android-App/src/androidTest/java/helper/RecyclerViewAssertions.java
@@ -1,9 +1,9 @@
-package de.luhmer.owncloudnewsreader.tests;
+package helper;
 
-import android.support.test.espresso.NoMatchingViewException;
-import android.support.test.espresso.ViewAssertion;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
+import androidx.test.espresso.NoMatchingViewException;
+import androidx.test.espresso.ViewAssertion;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.View;
 
 public class RecyclerViewAssertions implements ViewAssertion {