+++ News-Android-App/src/androidTest/java/screengrab/ScreenshotTest.java
@@ -2,9 +2,9 @@ package screengrab;
 
 import android.content.SharedPreferences;
 import android.preference.PreferenceManager;
-import android.support.test.rule.ActivityTestRule;
-import android.support.test.runner.AndroidJUnit4;
-import android.support.v4.view.GravityCompat;
+import androidx.test.rule.ActivityTestRule;
+import androidx.test.runner.AndroidJUnit4;
+import androidx.core.view.GravityCompat;
 
 import org.junit.Before;
 import org.junit.ClassRule;
@@ -31,8 +31,10 @@ import tools.fastlane.screengrab.locale.LocaleTestRule;
  */
 @RunWith(AndroidJUnit4.class)
 public class ScreenshotTest {
+
     @ClassRule
     public static final LocaleTestRule localTestRule = new LocaleTestRule();
+
     @Rule
     public ActivityTestRule<NewsReaderListActivity> mActivityRule = new ActivityTestRule<>(NewsReaderListActivity.class);