+++ News-Android-App/src/main/java/de/luhmer/owncloudnewsreader/adapter/DividerItemDecoration.java
@@ -5,9 +5,10 @@ import android.content.res.TypedArray;
 import android.graphics.Canvas;
 import android.graphics.Rect;
 import android.graphics.drawable.Drawable;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
 import android.view.View;
+
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 
 public class DividerItemDecoration extends RecyclerView.ItemDecoration {
 
@@ -40,7 +41,7 @@ public class DividerItemDecoration extends RecyclerView.ItemDecoration {
             return;
 
         if (!(parent.getLayoutManager() instanceof LinearLayoutManager) ||
-                ((LinearLayoutManager)parent.getLayoutManager()).getOrientation() != LinearLayoutManager.VERTICAL) {
+                ((LinearLayoutManager)parent.getLayoutManager()).getOrientation() != RecyclerView.VERTICAL) {
             throw new IllegalStateException(
                     "DividerItemDecoration can only be used with a vertical LinearLayoutManager.");
         }