+++ News-Android-App/src/main/java/de/luhmer/owncloudnewsreader/services/SyncItemStateService.java
@@ -25,7 +25,7 @@ import android.app.ActivityManager;
 import android.app.ActivityManager.RunningServiceInfo;
 import android.content.Context;
 import android.content.Intent;
-import android.support.v4.app.JobIntentService;
+import androidx.core.app.JobIntentService;
 import android.util.Log;
 
 import java.io.IOException;