+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 - Single Sign on for all Nextcloud Android Apps!
 
 
+0.9.9.26
+---------------------
+- Fix - <a href="https://github.com/owncloud/News-Android-App/issues/726">#726 Add new feed fails</a>
+- Fix - <a href="https://github.com/owncloud/News-Android-App/issues/744">#744 Fix issues when adding feeds (Thanks @Unpublished)</a>
+- Feature - <a href="https://github.com/owncloud/News-Android-App/issues/747">#747 Add option to share article when using chrome-custom-tabs</a>
+
+
 0.9.9.25
 ---------------------
 - Fix - app crashes